pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "code-intel"
version = "1.0.0"
description = "Code Usage Intelligence System — detect dead code and sensitive data leaks in CI/CD"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Code Intel Team"}
]
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "prometheus-client>=0.19.0",
    "Jinja2>=3.1.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
code-intel = "code_intel.cli:main"

[tool.setuptools.packages.find]
include = ["code_intel*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
